+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 # Change Log
 
-## NetKet 3.4 (⚙️ In development)
+## NetKet 3.3.1 (⚙️ In development)
 
 [GitHub commits](https://github.com/netket/netket/compare/v3.3...master).
 
@@ -15,7 +15,7 @@
 
 
 ### Bug Fixes
-
+* Initialisation of all implementations of `DenseSymm`, `DenseEquivariant`, `GCNN` now defaults to truncated normals with Lecun variance scaling. For layers without masking, there should be no noticeable change in behaviour. For masked layers, the same variance scaling now works correctly. [#1045](https://github.com/netket/netket/pull/1045)
 
 
 ## NetKet 3.3 (🎁 20 December 2021)